progress

+++ include/c/c_lib.h
@@ -6,8 +6,8 @@
 
 namespace cLib {
 
-s32 targetAngleY(const mVec3_c &target, const mVec3_c &source);
-s32 targetAngleX(const mVec3_c &target, const mVec3_c &source);
+s16 targetAngleY(const mVec3_c &target, const mVec3_c &source);
+s16 targetAngleX(const mVec3_c &target, const mVec3_c &source);
 void addCalcPos2(mVec3_c *src, const mVec3_c &target, f32 scale, f32 maxStep);
 f32 addCalcPosXZ(mVec3_c *src, const mVec3_c &target, f32 scale, f32 maxStep, f32 minStep);
 f32 addCalcPos(mVec3_c *src, const mVec3_c &target, f32 scale, f32 maxStep, f32 minStep);
